backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
python-telegram-bot==20.7
youtube-transcript-api==0.6.2
pytube==15.0.0
requests==2.31.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
openai==1.30.1
tiktoken==0.7.0
langdetect==1.0.9
gunicorn==22.0.0
eventlet==0.35.2
yt-dlp==2024.5.27
# ── Semantic retrieval upgrades ──────────────────────────────
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy>=1.24.0,<2